fixes some TOC issues, creates others
Without anything here, the table of contents in the pdf looks (visually) correct With \chapterstyle here, the chapter titles from this point forward are bold. In both cases, the numbering is correct. However, the navigation pane for the pdf does not nest correctly: chapter 2 is inside of section 1.3, part II is inside section 1.4, part V is inside chapter 11, etc.

Adding \sectionstyle here creates visual/numbering issues with the table of contents. However, in the navigation pane, the sections are all now properly nested.